requirements.txt
Django==1.8
List==1.3.0
Pillow==2.8.1
distribute==0.6.24
django-bootstrap3==5.3.1
django-crontab==0.6.0
gunicorn==19.3.0
pyserial==2.7
rrdtool==0.1.1
setproctitle==1.1.8
wsgiref==0.1.2
